requirements.txt
requests~=2.32.3
boto3~=1.38.33
Flask~=3.1.1
python-dotenv~=1.1.0
python-dateutil~=2.9.0.post0
twilio~=9.6.2